# Precompile most of the regular expressions